envs/data_pipeline_debug_env/server/requirements.txt
openenv-core[core]>=0.2.2
pydantic>=2.0
